openssl_1.0.2p: drop dependency on hostperl-runtime-native
In OE 2.1 and later, host perl is called "hostperl-runtime-native"
but prior to that host perl was called "perl-native-runtime" and
therefore a dependency on hostperl-runtime-native will cause errors
for OE 2.0 and before. See:

  http://git.openembedded.org/openembedded-core/commit/?id=f4dade8e765a8c7bfd131728b9e0a34631e24950

Although it would be possible to update the openssl recipe to
automatically detect the rename and depend on the appropriate recipe
name, for example:

  DEPENDS = "makedepend-native ${@bb.utils.contains('ASSUME_PROVIDED', 'perl-native-runtime', 'perl-native-runtime', 'hostperl-runtime-native', d)}"

it's not really necessary to do so since the host perl is always
included in ASSUME_PROVIDED (ie it's a dependency which is ignored by
bitbake and is mainly included in recipes to document the dependency
rather than for function reasons). The simplest way to allow the new
openssl recipe to be compatible with all versions of OE is to drop
the dependency on hostperl-runtime-native.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>